pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cubo"
version = "1.2.0"
description = "A modular Retrieval-Augmented Generation (RAG) system with FastAPI backend and Next.js frontend"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Paolo Astrino", email = "paolo@example.com"}
]
maintainers = [
    {name = "Paolo Astrino", email = "paolo@example.com"}
]
keywords = [
    "rag",
    "retrieval-augmented-generation",
    "nlp",
    "llm",
    "embedding",
    "semantic-search",
    "document-qa",
    "fastapi",
    "ollama"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "sentence-transformers>=3.0.0,<6.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.46.0,<5.0.0",
    "ollama>=0.3.0",
    "colorama>=0.4.0",
    "python-docx>=1.0.0",
    "PyPDF2>=3.0.0",
    "tokenizers>=0.15.0",
    "numpy>=1.21.0,<2.0.0",
    "protobuf>=3.20.0,<5.0.0",
    "cryptography>=40.0.0",
    "psutil>=5.9.0",
    "pandas>=1.3.0",
    "google-generativeai>=0.8.0",
    "pdf2image>=1.17.0",
    "Pillow>=10.0.0",
    "pyarrow>=13.0.0,<15.0.0",
    "openpyxl>=3.1.0",
    "pdfplumber>=0.10.3",
    "tqdm>=4.64.0",
    "datasketch",
    "networkx",
    "structlog",
    "python-json-logger",
    "whoosh",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "scikit-learn>=1.3.0",  # For semantic clustering in scaffolds
    "pytesseract>=0.3.10",  # For OCR (offline scanned PDF processing)
    "pdf2image>=1.16.3",  # For converting PDF pages to images for OCR
]

[project.optional-dependencies]
# GUI dependencies
gui = [
    "PySide6>=6.5.0",
]

# Dense retrieval with FAISS
faiss = [
    "faiss-cpu==1.7.4",
]

# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "reportlab>=4.0.0",  # For PDF generation in tests
]

# All optional dependencies combined
all = [
    "cubo[gui,faiss,dev]",
]

[project.urls]
Homepage = "https://github.com/PaoloAstrino/cubo"
Documentation = "https://github.com/PaoloAstrino/cubo#readme"
Repository = "https://github.com/PaoloAstrino/cubo"
"Bug Tracker" = "https://github.com/PaoloAstrino/cubo/issues"

[project.scripts]
cubo = "cubo.main:main"
cubo-server = "cubo.server.run:main"
cubo-gui = "cubo.gui.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["cubo*"]
namespaces = false

[tool.setuptools.package-data]
cubo = ["py.typed", "*.json", "*.yaml", "*.yml"]

# Black configuration
[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.eggs
  | build
  | dist
  | __pycache__
)/
'''

# Ruff configuration (modern Python linter)
[tool.ruff]
line-length = 100
target-version = "py38"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

# MyPy configuration
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true  # Enforce type hints on all functions
ignore_missing_imports = true
exclude = [
    "build",
    "dist",
    ".venv",
]

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests", "integration_tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
]
markers = [
    "requires_faiss: marks tests that require FAISS installation",
    "requires_whoosh: marks tests that require Whoosh installation",
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "e2e: marks tests as end-to-end tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# Coverage configuration
[tool.coverage.run]
source = ["src/cubo"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# Bandit security scanner configuration
[tool.bandit]
exclude_dirs = ["tests", "integration_tests", ".venv", "build", "dist"]
skips = ["B101", "B601"]  # Skip assert_used and paramiko calls
